Move MenuItemCompat into core and deprecate
The APIs it was adapting are now available in our minSdkVersion.
